spacesaver-transcode/app/requirements.txt
flask>=3.0
ffmpeg-python>=0.2
sqlmodel
fastapi
sqlalchemy
colorlog
